# worker/openapi.json
{
  "components": {
    "schemas": {
      "BlobInfo": {
        "properties": {
          "key": {
            "type": "string"
          },
          "sha256": {
            "type": "string"
          },
          "size": {
            "type": "integer"
          }
        },
        "required": [
          "key",
          "size"
        ],
        "type": "object"
      },
      "Error": {
        "$ref": "#/components/schemas/Error"
      },
...
      "GoldenList": {
        "properties": {
          "cursor": {
            "type": "string"
          },
          "objects": {
            "items": {
              "$ref": "#/components/schemas/BlobInfo"
            },
            "type": "array"
          }
        },
        "required": [
          "objects",
          "cursor"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "glaze-status-write": {
        "description": "CI's conformance job posting a glaze run. The Worker's secret GLAZE_STATUS_TOKEN; clients read it from GLAZE_STATUS_TOKEN.",
        "scheme": "bearer",
        "type": "http"
      },
      "golden-read": {
        "description": "reading the golden image: every machine that pulls. The Worker's secret GOLDEN_TOKEN; clients read it from IRGO_GOLDEN_TOKEN.",
        "scheme": "bearer",
        "type": "http"
      },
      "golden-write": {
        "description": "writing, deleting and listing it: only where you push. The Worker's secret GOLDEN_PUSH_TOKEN; clients read it from IRGO_GOLDEN_PUSH_TOKEN.",
        "scheme": "bearer",
        "type": "http"
      },
      "jobs": {
        "description": "a remote job's caller: submits and reads its own jobs (the secret is name=token,name=token, one per caller). The Worker's secret JOBS_TOKENS; clients read it from IRGO_REMOTE_TOKEN.",
        "scheme": "bearer",
        "type": "http"
      },
      "jobs-admin": {
        "description": "lists every job in the queue. The Worker's secret JOBS_ADMIN_TOKEN; clients read it from IRGO_REMOTE_ADMIN_TOKEN.",
        "scheme": "bearer",
        "type": "http"
      },
      "jobs-runner": {
        "description": "the Mac running irgo-winvm serve: takes jobs and reports them. The Worker's secret JOBS_RUNNER_TOKEN; clients read it from IRGO_REMOTE_RUNNER_TOKEN.",
        "scheme": "bearer",
        "type": "http"
      },
      "ledger-read": {
        "description": "reading the ledger: its page and its JSON. The Worker's secret LEDGER_READ_TOKEN; clients read it from IRGO_LEDGER_READ_TOKEN.",
        "scheme": "bearer",
        "type": "http"
      },
      "ledger-write": {
        "description": "the tool posting events to the ledger. The Worker's secret LEDGER_TOKEN; clients read it from IRGO_LEDGER_TOKEN.",
        "scheme": "bearer",
        "type": "http"
      }
...
  },
  "info": {
...
    "title": "irgo-windows-vm Worker",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/golden-list/{kind}": {
      "get": {
        "operationId": "golden-list",
        "parameters": [
          {
            "description": "manifests or chunks",
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "the previous page's cursor; empty for the first page",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoldenList"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code not-found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code internal"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code storage"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code not-configured"
          }
        },
        "security": [
          {
            "golden-write": []
          }
        ],
        "summary": "one page of the manifests' or the chunks' keys and sizes, for vm-golden-push -delete, which must find every chunk no manifest names. The write token, so a machine that only pulls cannot enumerate the bucket",
        "x-irgo-winvm-commands": [
          "vm-golden-push"
        ],
        "x-scope": "golden-write"
      }
    },
    "/api/golden/{key}": {
      "delete": {
        "operationId": "golden-delete",
        "parameters": [
          {
            "description": "golden/latest, golden/manifests/\u003csha256\u003e.json or golden/chunks/\u003csha256\u003e.zst; anything else is 404 (spans path segments: its slashes are not escaped)",
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code not-found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code internal"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code storage"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code not-configured"
          }
        },
        "security": [
          {
            "golden-write": []
          }
        ],
        "summary": "remove an object; nothing there is success too",
        "x-irgo-winvm-commands": [
          "vm-golden-push"
        ],
        "x-scope": "golden-write"
      },
      "get": {
        "operationId": "golden-get",
        "parameters": [
          {
            "description": "golden/latest, golden/manifests/\u003csha256\u003e.json or golden/chunks/\u003csha256\u003e.zst; anything else is 404 (spans path segments: its slashes are not escaped)",
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "optional, one range: bytes=\u003cfirst\u003e-[\u003clast\u003e] or bytes=-\u003cn\u003e",
            "in": "header",
            "name": "Range",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "contentMediaType": "application/octet-stream",
                  "type": "string"
                }
              }
            },
            "description": "OK",
            "headers": {
              "X-Golden-Sha256": {
                "description": "its SHA-256 as verified when it was stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Golden-Size": {
                "description": "the whole object's size",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "206": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "contentMediaType": "application/octet-stream",
                  "type": "string"
                }
              }
            },
            "description": "Partial Content",
            "headers": {
              "X-Golden-Sha256": {
                "description": "its SHA-256 as verified when it was stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Golden-Size": {
                "description": "the whole object's size",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code not-found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code conflict"
          },
          "416": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code range-not-satisfiable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code internal"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code storage"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code not-configured"
          }
        },
        "security": [
          {
            "golden-read": []
          }
        ],
        "summary": "the object, streamed from R2 without passing through Go; a single byte Range answers 206, which is how a pull resumes",
        "x-irgo-winvm-commands": [
          "vm-golden-pull",
          "vm-golden-push",
          "vm-create"
        ],
        "x-scope": "golden-read"
      },
      "head": {
        "operationId": "golden-head",
        "parameters": [
          {
            "description": "golden/latest, golden/manifests/\u003csha256\u003e.json or golden/chunks/\u003csha256\u003e.zst; anything else is 404 (spans path segments: its slashes are not escaped)",
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "X-Golden-Sha256": {
                "description": "its SHA-256 as verified when it was stored; absent for one stored another way",
                "schema": {
                  "type": "string"
                }
              },
              "X-Golden-Size": {
                "description": "the whole object's size (Workers drops Content-Length from an answer to HEAD)",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "code unauthorized"
          },
          "404": {
            "description": "code not-found"
          },
          "500": {
            "description": "code internal"
          },
          "502": {
            "description": "code storage"
          },
          "503": {
            "description": "code not-configured"
          }
        },
        "security": [
          {
            "golden-read": []
          }
        ],
        "summary": "an object's size and SHA-256, without its bytes; 404 with no body when it is not there",
        "x-irgo-winvm-commands": [
          "vm-golden-pull",
          "vm-golden-push",
          "vm-create"
        ],
        "x-scope": "golden-read"
      },
      "put": {
        "operationId": "golden-put",
        "parameters": [
          {
            "description": "golden/latest, golden/manifests/\u003csha256\u003e.json or golden/chunks/\u003csha256\u003e.zst; anything else is 404 (spans path segments: its slashes are not escaped)",
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "required: the body's SHA-256, 64 lower-case hex digits",
            "in": "header",
            "name": "X-Golden-Sha256",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlobInfo"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code bad-request or digest-mismatch"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code not-found"
          },
          "411": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code length-required"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code too-large"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code internal"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code storage"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code not-configured"
          }
        },
        "security": [
          {
            "golden-write": []
          }
        ],
        "summary": "store an object, only if it hashes to X-Golden-Sha256 (R2 checks it as the stream arrives); a manifest's claim must also be its name",
        "x-irgo-winvm-commands": [
          "vm-golden-push"
        ],
        "x-max-body": 83886080,
        "x-scope": "golden-write"
      }
    },
...
  }
}